Extract Tags struct from table_manager.go
Seemed odd to have schema_config.go depend on table_manager.go, and it
works ok as a self-contained file.
